+++ src/app/(route)/page.tsx
@@ -1,5 +1,12 @@
+import DDuDuSheet from "../_components/client/DDuDuSheet/DDuDuSheet";
+
 const Home = () => {
-  return <main className="h-full w-full">DDuDu</main>;
+  return (
+    <main className="h-full w-full">
+      DDuDu
+      <DDuDuSheet />
+    </main>
+  );
 };
 
 export default Home;
