+++ src/app/_components/server/index.ts
@@ -1,4 +1,4 @@
 export * from "./icons";
 
 export { default as ExampleIcon } from "./ExampleIcon/ExampleIcon";
-export { default as DDuDuReaptItem } from "./DDuDuReaptItem/DDuDuReaptItem";
+export { default as GoalTodoListItem } from "./GoalTodoListItem/GoalTodoListItem";
