--- src/stories/components/DduduReaptItem_Storybook/components/DDuDuReaptItemView/DDuDuReaptItemView.tsx
@@ -1,8 +1,0 @@
-import { DDuDuReaptItem } from "@/app/_components/server";
-import { DDuDuReaptItemProps } from "@/app/_components/server/DDuDuReaptItem/DDuDuReaptItem";
-
-const DDuDuReaptItemView = (props: DDuDuReaptItemProps) => {
-  return <DDuDuReaptItem {...props} />;
-};
-
-export default DDuDuReaptItemView;
