--- src/stories/components/DduduReaptItem_Storybook/components/index.ts
@@ -1,1 +1,0 @@
-export { default as DDuDuReaptItemView } from "./DDuDuReaptItemView/DDuDuReaptItemView";
