+++ src/stories/components/GoalTodoListItem_Storybook/components/GoalTodoListItem/GoalTodoListItemView.tsx
@@ -1,0 +1,8 @@
+import { GoalTodoListItem } from "@/app/_components/server";
+import { GoalTodoListItemProps } from "@/app/_components/server/GoalTodoListItem/GoalTodoListItem";
+
+const GoalTodoListItemView = (props: GoalTodoListItemProps) => {
+  return <GoalTodoListItem {...props} />;
+};
+
+export default GoalTodoListItemView;
