+++ src/stories/components/GoalTodoListItem_Storybook/components/index.ts
@@ -1,0 +1,1 @@
+export { default as GoalTodoListItemView } from "./GoalTodoListItem/GoalTodoListItemView";
